Projects/meme_scout/requirements.txt
tweepy==4.14.0
pandas==2.0.3
python-dotenv==1.0.0
